requirements.txt
bleak>=0.21.0
python-dotenv==1.0.1
